requirements.txt
python-telegram-bot==20.7
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
psutil==5.9.6 